+++ src/main/kotlin/com/highv/ecommerce/domain/backoffice/controller/SalesStatisticsController.kt
@@ -1,8 +1,8 @@
 package com.highv.ecommerce.domain.backoffice.controller
 
-import com.highv.ecommerce.domain.backoffice.dto.ProductQuantityResponse
-import com.highv.ecommerce.domain.backoffice.dto.ProductSalesResponse
-import com.highv.ecommerce.domain.backoffice.dto.TotalSalesResponse
+import com.highv.ecommerce.domain.backoffice.dto.salesstatics.ProductQuantityResponse
+import com.highv.ecommerce.domain.backoffice.dto.salesstatics.ProductSalesResponse
+import com.highv.ecommerce.domain.backoffice.dto.salesstatics.TotalSalesResponse
 import com.highv.ecommerce.domain.backoffice.service.SalesStatisticsService
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
