--- src/main/kotlin/com/highv/ecommerce/domain/backoffice/dto/TotalSalesResponse.kt
@@ -1,5 +1,0 @@
-package com.highv.ecommerce.domain.backoffice.dto
-
-data class TotalSalesResponse(
-    val totalPrice: Int
-)
