+++ src/main/kotlin/com/highv/ecommerce/domain/backoffice/dto/salesstatics/ProductSalesResponse.kt
@@ -1,4 +1,4 @@
-package com.highv.ecommerce.domain.backoffice.dto
+package com.highv.ecommerce.domain.backoffice.dto.salesstatics
 
 data class ProductSalesResponse(
     val productName: String,
