+++ src/main/kotlin/com/highv/ecommerce/domain/backoffice/dto/salesstatics/TotalSalesResponse.kt
@@ -1,0 +1,5 @@
+package com.highv.ecommerce.domain.backoffice.dto.salesstatics
+
+data class TotalSalesResponse(
+    val totalPrice: Int
+)
